FEAT: move 'CloudEventBackgroundJob' to sep project
