+++ src/Arcus.BackgroundJobs/Arcus.BackgroundJobs.csproj
@@ -6,7 +6,7 @@
     <Company>Arcus</Company>
     <RepositoryType>Git</RepositoryType>
     <PackageTags>Azure;WebAPI;App Services;Web App;Web;API</PackageTags>
-    <Description>Provides capabilities for running background jobs to automate workflows such as updating cached Azure Key Value secrets.</Description>
+    <Description>Provides capabilities for running background jobs to automate workflows.</Description>
     <Copyright>Copyright (c) Arcus</Copyright>
     <PackageLicenseUrl>https://github.com/arcus-azure/arcus.webapi/blob/master/LICENSE</PackageLicenseUrl>
     <PackageProjectUrl>https://github.com/arcus-azure/arcus.webapi</PackageProjectUrl>
@@ -19,14 +19,7 @@
   <ItemGroup>
     <PackageReference Include="Arcus.EventGrid" Version="3.0.0-preview-1" />
     <PackageReference Include="Arcus.Messaging.Pumps.ServiceBus" Version="20200115.0.0-PR-32" />
-    <PackageReference Include="Arcus.Security.Core" Version="20200107.0.0-PR-120" />
-    <PackageReference Include="Arcus.Security.Providers.AzureKeyVault" Version="1.0.0" />
     <PackageReference Include="Microsoft.Azure.EventGrid" Version="3.2.0" />
-    <PackageReference Include="Microsoft.Azure.KeyVault.Core" Version="3.0.4" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\Arcus.BackgroundJobs\Arcus.BackgroundJobs.csproj" />
   </ItemGroup>
 
   <PropertyGroup>
