+++ src/Arcus.WebApi.sln
@@ -19,7 +19,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Security.Authe
 EndProject
 Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Jobs", "Arcus.WebApi.Jobs\Arcus.WebApi.Jobs.csproj", "{4062664C-2421-45D7-A986-E0C401BCDFA0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Arcus.WebApi.Integration", "Arcus.WebApi.Integration\Arcus.WebApi.Integration.csproj", "{257F0347-4F29-4386-B2A2-DEBF947C6E41}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "Arcus.WebApi.Integration", "Arcus.WebApi.Integration\Arcus.WebApi.Integration.csproj", "{257F0347-4F29-4386-B2A2-DEBF947C6E41}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Arcus.BackgroundJobs", "Arcus.BackgroundJobs\Arcus.BackgroundJobs.csproj", "{920EDCF7-74A3-474C-A412-58D511FFC63C}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -59,6 +61,10 @@ Global
 		{257F0347-4F29-4386-B2A2-DEBF947C6E41}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{257F0347-4F29-4386-B2A2-DEBF947C6E41}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{257F0347-4F29-4386-B2A2-DEBF947C6E41}.Release|Any CPU.Build.0 = Release|Any CPU
+		{920EDCF7-74A3-474C-A412-58D511FFC63C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{920EDCF7-74A3-474C-A412-58D511FFC63C}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{920EDCF7-74A3-474C-A412-58D511FFC63C}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{920EDCF7-74A3-474C-A412-58D511FFC63C}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
